requirements.txt
transformers
pillow
streamlit
torch
requests
openai
python-dotenv